Move FAB initialization to the FAB provider
This moves the fab initialization codes to the FAB provider so we can have fewer FAB codes in core Airflow.
This is in line with the preparation to remove FAB as a core dependency.
