+++ airflow/providers/fab/webserver/init_views.py
@@ -1,0 +1,97 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+from __future__ import annotations
+
+from airflow.security import permissions
+
+
+def init_appbuilder_views(app):
+    """Initialize Web UI views."""
+    from airflow.www import views
+
+    appbuilder = app.appbuilder
+
+    # Remove the session from scoped_session registry to avoid
+    # reusing a session with a disconnected connection
+    appbuilder.session.remove()
+    appbuilder.add_view_no_menu(views.AutocompleteView())
+    appbuilder.add_view_no_menu(views.Airflow())
+    appbuilder.add_view(
+        views.DagRunModelView,
+        permissions.RESOURCE_DAG_RUN,
+        category=permissions.RESOURCE_BROWSE_MENU,
+        category_icon="fa-globe",
+    )
+    appbuilder.add_view(
+        views.JobModelView, permissions.RESOURCE_JOB, category=permissions.RESOURCE_BROWSE_MENU
+    )
+    appbuilder.add_view(
+        views.LogModelView, permissions.RESOURCE_AUDIT_LOG, category=permissions.RESOURCE_BROWSE_MENU
+    )
+    appbuilder.add_view(
+        views.VariableModelView, permissions.RESOURCE_VARIABLE, category=permissions.RESOURCE_ADMIN_MENU
+    )
+    appbuilder.add_view(
+        views.TaskInstanceModelView,
+        permissions.RESOURCE_TASK_INSTANCE,
+        category=permissions.RESOURCE_BROWSE_MENU,
+    )
+    appbuilder.add_view(
+        views.TaskRescheduleModelView,
+        permissions.RESOURCE_TASK_RESCHEDULE,
+        category=permissions.RESOURCE_BROWSE_MENU,
+    )
+    appbuilder.add_view(
+        views.TriggerModelView,
+        permissions.RESOURCE_TRIGGER,
+        category=permissions.RESOURCE_BROWSE_MENU,
+    )
+    appbuilder.add_view(
+        views.ConfigurationView,
+        permissions.RESOURCE_CONFIG,
+        category=permissions.RESOURCE_ADMIN_MENU,
+        category_icon="fa-user",
+    )
+    appbuilder.add_view(
+        views.ConnectionModelView, permissions.RESOURCE_CONNECTION, category=permissions.RESOURCE_ADMIN_MENU
+    )
+    appbuilder.add_view(
+        views.SlaMissModelView, permissions.RESOURCE_SLA_MISS, category=permissions.RESOURCE_BROWSE_MENU
+    )
+    appbuilder.add_view(
+        views.PluginView, permissions.RESOURCE_PLUGIN, category=permissions.RESOURCE_ADMIN_MENU
+    )
+    appbuilder.add_view(
+        views.ProviderView, permissions.RESOURCE_PROVIDER, category=permissions.RESOURCE_ADMIN_MENU
+    )
+    appbuilder.add_view(
+        views.PoolModelView, permissions.RESOURCE_POOL, category=permissions.RESOURCE_ADMIN_MENU
+    )
+    appbuilder.add_view(
+        views.XComModelView, permissions.RESOURCE_XCOM, category=permissions.RESOURCE_ADMIN_MENU
+    )
+    appbuilder.add_view(
+        views.DagDependenciesView,
+        permissions.RESOURCE_DAG_DEPENDENCIES,
+        category=permissions.RESOURCE_BROWSE_MENU,
+    )
+    # add_view_no_menu to change item position.
+    # I added link in extensions.init_appbuilder_links.init_appbuilder_links
+    appbuilder.add_view_no_menu(views.RedocView)
+    # Development views
+    appbuilder.add_view_no_menu(views.DevView)
+    appbuilder.add_view_no_menu(views.DocsView)
