+++ airflow/www/extensions/init_views.py
@@ -28,7 +28,6 @@ from flask import request
 
 from airflow.api_connexion.exceptions import common_error_handler
 from airflow.configuration import conf
-from airflow.security import permissions
 from airflow.utils.yaml import safe_load
 from airflow.www.constants import SWAGGER_BUNDLE, SWAGGER_ENABLED
 from airflow.www.extensions.init_auth_manager import get_auth_manager
@@ -47,88 +46,6 @@ def init_flash_views(app):
     from airflow.www.blueprints import routes
 
     app.register_blueprint(routes)
-
-
-def init_appbuilder_views(app):
-    """Initialize Web UI views."""
-    from airflow.models import import_all_models
-
-    import_all_models()
-
-    from airflow.www import views
-
-    appbuilder = app.appbuilder
-
-    # Remove the session from scoped_session registry to avoid
-    # reusing a session with a disconnected connection
-    appbuilder.session.remove()
-    appbuilder.add_view_no_menu(views.AutocompleteView())
-    appbuilder.add_view_no_menu(views.Airflow())
-    appbuilder.add_view(
-        views.DagRunModelView,
-        permissions.RESOURCE_DAG_RUN,
-        category=permissions.RESOURCE_BROWSE_MENU,
-        category_icon="fa-globe",
-    )
-    appbuilder.add_view(
-        views.JobModelView, permissions.RESOURCE_JOB, category=permissions.RESOURCE_BROWSE_MENU
-    )
-    appbuilder.add_view(
-        views.LogModelView, permissions.RESOURCE_AUDIT_LOG, category=permissions.RESOURCE_BROWSE_MENU
-    )
-    appbuilder.add_view(
-        views.VariableModelView, permissions.RESOURCE_VARIABLE, category=permissions.RESOURCE_ADMIN_MENU
-    )
-    appbuilder.add_view(
-        views.TaskInstanceModelView,
-        permissions.RESOURCE_TASK_INSTANCE,
-        category=permissions.RESOURCE_BROWSE_MENU,
-    )
-    appbuilder.add_view(
-        views.TaskRescheduleModelView,
-        permissions.RESOURCE_TASK_RESCHEDULE,
-        category=permissions.RESOURCE_BROWSE_MENU,
-    )
-    appbuilder.add_view(
-        views.TriggerModelView,
-        permissions.RESOURCE_TRIGGER,
-        category=permissions.RESOURCE_BROWSE_MENU,
-    )
-    appbuilder.add_view(
-        views.ConfigurationView,
-        permissions.RESOURCE_CONFIG,
-        category=permissions.RESOURCE_ADMIN_MENU,
-        category_icon="fa-user",
-    )
-    appbuilder.add_view(
-        views.ConnectionModelView, permissions.RESOURCE_CONNECTION, category=permissions.RESOURCE_ADMIN_MENU
-    )
-    appbuilder.add_view(
-        views.SlaMissModelView, permissions.RESOURCE_SLA_MISS, category=permissions.RESOURCE_BROWSE_MENU
-    )
-    appbuilder.add_view(
-        views.PluginView, permissions.RESOURCE_PLUGIN, category=permissions.RESOURCE_ADMIN_MENU
-    )
-    appbuilder.add_view(
-        views.ProviderView, permissions.RESOURCE_PROVIDER, category=permissions.RESOURCE_ADMIN_MENU
-    )
-    appbuilder.add_view(
-        views.PoolModelView, permissions.RESOURCE_POOL, category=permissions.RESOURCE_ADMIN_MENU
-    )
-    appbuilder.add_view(
-        views.XComModelView, permissions.RESOURCE_XCOM, category=permissions.RESOURCE_ADMIN_MENU
-    )
-    appbuilder.add_view(
-        views.DagDependenciesView,
-        permissions.RESOURCE_DAG_DEPENDENCIES,
-        category=permissions.RESOURCE_BROWSE_MENU,
-    )
-    # add_view_no_menu to change item position.
-    # I added link in extensions.init_appbuilder_links.init_appbuilder_links
-    appbuilder.add_view_no_menu(views.RedocView)
-    # Development views
-    appbuilder.add_view_no_menu(views.DevView)
-    appbuilder.add_view_no_menu(views.DocsView)
 
 
 def init_plugins(app):
