+++ tests/providers/fab/auth_manager/conftest.py
@@ -27,7 +27,7 @@ from tests.test_utils.decorators import dont_initialize_flask_app_submodules
 def minimal_app_for_auth_api():
     @dont_initialize_flask_app_submodules(
         skip_all_except=[
-            "init_appbuilder",
+            "init_fab",
             "init_api_auth",
             "init_api_auth_provider",
             "init_api_error_handlers",
