+++ tests/www/test_app.py
@@ -260,7 +260,7 @@ class TestApp:
 
 
 class TestFlaskCli:
-    @dont_initialize_flask_app_submodules(skip_all_except=["init_appbuilder"])
+    @dont_initialize_flask_app_submodules(skip_all_except=["init_fab"])
     def test_flask_cli_should_display_routes(self, capsys):
         with mock.patch.dict("os.environ", FLASK_APP="airflow.www.app:cached_app"), mock.patch.object(
             sys, "argv", ["flask", "routes"]
