+++ tests/www/views/conftest.py
@@ -52,9 +52,7 @@ def app(examples_dag_bag):
     @dont_initialize_flask_app_submodules(
         skip_all_except=[
             "init_api_connexion",
-            "init_appbuilder",
-            "init_appbuilder_links",
-            "init_appbuilder_views",
+            "init_fab",
             "init_flash_views",
             "init_jinja_globals",
             "init_plugins",
