+++ tests/www/views/test_views_rate_limit.py
@@ -32,9 +32,7 @@ def app_with_rate_limit_one(examples_dag_bag):
     @dont_initialize_flask_app_submodules(
         skip_all_except=[
             "init_api_connexion",
-            "init_appbuilder",
-            "init_appbuilder_links",
-            "init_appbuilder_views",
+            "init_fab",
             "init_flash_views",
             "init_jinja_globals",
             "init_plugins",
